Implement doubleEveryOther
This function doubles every other integer in a list starting from the
right. So it doubles the second to last element in a list and then goes
back from there.

The two `reverse` things are _dumb_ and I hate them, but I'm unclear on
how else to get this to work, and it's a good enough first pass.