Improve style for toDigits
This used to be a pointfree function, which was super nice. Then I had
to check the value of `n` to make sure if was positive, and that screwed
everything up. To get it to compile, I just tossed an `$` at the end,
but it actually reads better if we use function application _instead_ of
function composition.